Add tests to all handlers.
